feat(pos): add "Copy entry" button to log cards
- Add a compact "Copy entry" pill to each log card header; tapping it copies
  the entry to the clipboard as structured JSON (type/date/location/message/
  body, full date with year) and shows a confirmation toast.
- Add an optional compact size to the shared Button (size="sm"); defaults to
  the existing size so other call sites are unaffected.
- Right-align the copy button by lifting the card header to a full-width row.
- Extract log helpers to utils/logs.ts (formatTimestamp, formatFullTimestamp,
  buildLogText) and the card into components/log-card.tsx (memoized), slimming
  app/logs.tsx to the screen concern.
- Give the Logs header title component a display name (react/display-name).
